refactor(/model, /test): model 이하의 클래스들을 game 과 car 패키지로 분류해라
