+++ src/main/java/racingcar/model/car/CarStatus.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.car;
 
 public record CarStatus(String name, int position) {
 }
